% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeDim}
\alias{makeDim}
\title{Make dictionary dimensions}
\usage{
makeDim(DT = NULL)
}
\arguments{
\item{DT}{data.table a fact table.}
}
\value{
data.table fact table including time and/or geography dimension (depending on the input columns).
}
\description{
Lookup basic time and geography dimensions. You can use it on your own dataset. Be aware you need to have \code{date # Date type} and \code{country # character type} columns in fact table \code{DT}.
}
\note{
Required fields \code{date} and \code{country} can be also data types which can be directly converted using \code{as.character} and \code{as.Date}. See references for geography dim key details.
}
\examples{
\dontrun{
  # load shinyBI sample DT
  DT <- readRDS(system.file("extdata","cran_logs.rds",package="shinyBI"))
  DT <- makeDim(DT)
  shinyBI()

  # load dataset from csv
  DT <- fread("source.csv")
  DT <- makeDim(DT)
  shinyBI()
}
}
\references{
\url{http://dev.maxmind.com/geoip/legacy/codes/iso3166/}
}

