% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeGeographyDim}
\alias{makeGeographyDim}
\title{makeGeographyDim}
\usage{
makeGeographyDim(DT)
}
\arguments{
\item{DT}{data.table a fact table.}
}
\description{
Basic geography dimension on \code{country} (\code{character}) column in \code{DT}. Use \code{\link{makeDim}} function instead.
}

