% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeTimeDim}
\alias{makeTimeDim}
\title{makeTimeDim}
\usage{
makeTimeDim(DT)
}
\arguments{
\item{DT}{data.table a fact table.}
}
\description{
Basic time dimension on \code{date} (\code{Date}) column in \code{DT}. Use \code{\link{makeDim}} function instead.
}

