% Generated by roxygen2 (4.0.1): do not edit by hand
\name{translation}
\alias{translation}
\title{Translations}
\usage{
translation(x)
}
\arguments{
\item{x}{character vector.}
}
\value{
named vector, user friendly names.
}
\description{
Format labels for UI objects.
}
\examples{
\dontrun{
  translation(c("year_week","year_month","downloads"))
}
}

